requirements.txt
dash==0.21.1
decorator==4.2.1
pipenv==11.10.1
plotly==2.5.0
numpy==1.14.2
PyVISA==1.9.0
serial==0.0.54
dash-renderer==0.12.1
dash-html-components==0.10.1
dash-core-components==0.22.1
gunicorn
dash_daq-0.0.2.tar.gz
pylint
flake8
blinkstick
selenium
percy